Update base58 encoding package to improve readability and performance
The base58 encoding package has been refactored with an aim to improve readability and efficiency as follows:
- The base58 table was changed from a hardcoded array to a function that dynamically initializes the table.
- The Encoding struct was introduced to make parameterizing the base58 alphabet possible.
- Helper functions like `tableInitializer()`, `appendZeroBytes()`, `reverseByteOrder()`, `calculateResult()`, `createTmpVariables()` and `countZeros()` were implemented for better code modularity and readability.
- Change made to `go.mod` to point to a new repository and upgrading the Go language version.

Also, unnecessary benchmark tests were deleted and README was updated to reflect the changes in package usage procedures. The `.gitignore` file was added to exclude unnecessary files and folders from the repository. These changes improve code maintainability and ease of use for developers.